GtkBox: Don't manually specify a padding
This might look good on my display, but on a display with a different
DPI value, it's probably garbage. Specifying space in pixels is
meaningless today.

Yes, this applies to the level bar and the window size as well. I have
not yet found a better way to control those, though.