Retab
For years, I've been using a tab size of 4. This, however, conflicts
with the idea of limiting the line length: You can only limit the line
length in a meaningful way if you're using the default tab size of 8.
But 8 is a huge waste of space...

So let's just do this. Retab to 4 spaces and limit the line length to
about 80 characters.