Remove GtkLevelBar, use bar included in GtkEntry
I was still seeing this warning whenever the GtkLevelBar was set to a
value of 0:

    Negative content width -2 (allocation 0, extents 1x1) while
    allocating gadget (node block, owner GtkLevelBar)

This happens even in a minimal test program:

    #include <gtk/gtk.h>

    int
    main(int argc, char **argv)
    {
        GtkWidget *win, *progress;

        gtk_init(&argc, &argv);

        win = gtk_window_new(GTK_WINDOW_TOPLEVEL);

        progress = gtk_level_bar_new();
        gtk_level_bar_set_value(GTK_LEVEL_BAR(progress), 0);

        gtk_container_add(GTK_CONTAINER(win), progress);
        gtk_widget_show_all(win);
        gtk_main();
    }

It would appear that it's illegal to call gtk_level_bar_set_value() with
a value of 0. Or, and that's just as likely, I don't understand how a
GtkLevelBar is supposed to work. You don't even have to call
gtk_level_bar_set_value() at all, since 0 is the default value of such a
bar.

All of this doesn't really matter, though, since GtkEntry has a built-in
progress bar that we can use.

CC #20.